Update to newer pandas version (#488)
Three were two notebooks that used `hpo_trained.summary().loss.argmax()`. In the older version of pandas, that returned the key into the dataframe index, which was the pipeline name in the summary. But with the newest pandas, it returns an integer instead. Fixed the notebooks so they work either way.